JavaScript calculator and changes to HTML and CSS
The JS calculator uses an external file containing a method to convert
numbers into their corresponding strings, yielding an output such as
“You need one 20lbs. bag”  Changes to HTML and CSS to make the page
more fluid.

Found a problem: occurred when the Bootstrap code for input with
imbedded lateral button was inserted to create the calculator layout.
The footer is now on top of the jumbotron.